Split Memory object into Buffer and Image
